Rename getCredentialFromFacebook services to be more unique
