+++ services/auth/getCredentialFromFacebook/logInToFacebook/index.js
@@ -1,6 +1,6 @@
 import { LoginManager } from 'react-native-fbsdk';
 
-export default function logIn() {
+export default function logInToFacebook() {
   return new Promise((resolve, reject) => {
     LoginManager.logInWithReadPermissions(['public_profile']).then(
       (result) => {
