+++ services/auth/getCredentialFromFacebook/logOutFromFacebook/__tests__/logOutFromFacebook.test.js
